Update Econia named address for mainnet dependency
Econia Move module source code in this commit is
identical to that of 8148afe (tag v4.0.2-audited),
which was published on mainnet in transaction
154118354.

However, that revision also includes a Move script
that may prevent other packages from using it as a
dependency in their Move.toml files. Hence this
commit is put forth for other projects to use as a
mainnet build dependency.